1、我解决了下载报错问题。它是由于uart.c部分的代码异常导致的，同时，不要使用printf函数打印任何特殊字符（中文也算） 2、encoder编码器驱动没有完成。其代码完成，具体问题查找需要借助示波器 3、完善了电机驱动部分的代码。新增motorspeed函数，可以更流畅地控制电机

+++ car_motor_control.h
@@ -33,7 +33,7 @@ void Motor_init(void);
  * @param leftSpeed  左轮速度 (0-100)
  * @param rightSpeed 右轮速度 (0-100)
  */
-void Motor_setSpeed(uint8_t leftSpeed, uint8_t rightSpeed);
+void Motor_setSpeed(int16_t leftSpeed, int16_t rightSpeed);
 
 /**
  * 小车前进
